Lower XPCMachServer-specific API to a subclass
